refactor(worker): extract workers/list-query.mjs from api.mjs (#510)
Second slice of the api.mjs de-monolith. Move the 9 list-query transform
helpers (applyQueryFilters + filterRows/applyListTransform/searchRows/sortRows/
compareValues/paginateRows/validateListQuery/integerParam) into a new leaf
module workers/list-query.mjs.

Pure restructuring — NO behaviour change. The module imports only
API_QUERY_COLLECTIONS from ../src/contracts.mjs and nothing from api.mjs, so
there is no cycle. applyQueryFilters is the single public export (one call site
in handleApiRequest); the other eight are module-private. Dropped the now-unused
API_QUERY_COLLECTIONS import from api.mjs.

api.mjs: 3526 -> 3301 LOC. Verified by dependency analysis: only applyQueryFilters
was referenced outside the block; no test imports these symbols by name (they're
exercised via handleRequest), so no re-exports needed.

Verified: 168 worker tests pass (api-coverage, worker-runtime, subnet-overview,
search-quality, subnet-slug, batch-ratelimit, ai-search), worker:test (esbuild)
green, validate:contract-drift clean, lint clean. No logic diff.
